+++ src/data/records/sam/std_traits.rs
@@ -116,3 +116,43 @@ impl std::fmt::Display for SamDataViewMut<'_> {
         )
     }
 }
+
+impl Display for SamOptField {
+    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
+        write!(f, "{}{}:{}", self.tag[0] as char, self.tag[1] as char, self.value)
+    }
+}
+
+impl Display for SamOptValue {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        match self {
+            SamOptValue::Char(val) => write!(f, "A:{val}", val = *val as char),
+            SamOptValue::Int(val) => write!(f, "i:{val}"),
+            SamOptValue::Float(val) => write!(f, "f:{val}"),
+            SamOptValue::String(val) => write!(f, "Z:{val}"),
+            SamOptValue::Hex(val) => write!(f, "H:{val}"),
+            SamOptValue::Array(val) => match val {
+                OptArray::I8(vec) => OptArray::fmt_opt_array(f, 'c', vec),
+                OptArray::U8(vec) => OptArray::fmt_opt_array(f, 'C', vec),
+                OptArray::I16(vec) => OptArray::fmt_opt_array(f, 's', vec),
+                OptArray::U16(vec) => OptArray::fmt_opt_array(f, 'S', vec),
+                OptArray::I32(vec) => OptArray::fmt_opt_array(f, 'i', vec),
+                OptArray::U32(vec) => OptArray::fmt_opt_array(f, 'I', vec),
+                OptArray::F32(vec) => OptArray::fmt_opt_array(f, 'f', vec),
+            },
+        }
+    }
+}
+
+impl OptArray {
+    /// A helper function for displaying a [`SamOptValue::Array`].
+    fn fmt_opt_array<T: Display>(f: &mut Formatter<'_>, arr_type: char, vals: &[T]) -> std::fmt::Result {
+        write!(f, "B:{arr_type}")?;
+
+        for v in vals {
+            write!(f, ",{v}")?;
+        }
+
+        Ok(())
+    }
+}